backend/backend_/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
langchain
langchain-community
langchain-google-genai
chromadb
python-dotenv
pandas
streamlit
plotly